Scaffolds the project
TL;DR
-----

Frames up the project

Details
-------

Prepares the baseline for the project. Includes basics like a Makefile
for packaging and releasing, a Brewfile for installing dependencies,
.envrc for configuring the environment, and a .gitignore that should
make sense of the project. Since this will be release with Replicated,
we also incoprorate the manifest that the Platform expects.

Also provides some baseline configuration for Aider, including some
coding conventions.